pyproject.toml
[tool.poetry]
name = "edunft"
version = "0.1.0"
description = ""
authors = ["7ongxia <7ongxia@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
mangum = "^0.17.0"
openai = "^0.26.0"
python-multipart = "^0.0.5"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

